Initial commit of my solution
First commit of my completed solution. Believed to be working correctly,
but to be double checked.
cachevector.R script is the example code provided in the assignment
brief (README.md).
cachematrix.R script is my solution to the assignment.
README.html is a HTML formatted version of the original MarkDown file,
for better readability.